Include cstdio so that this header file can compile with gcc 4.4.
Otherwise printf is undeclared. This code worked with earlier versions
of gcc because either cstdio or stdio.h was being included by some other
header file. But this was not guaranteed behavior, so with GCC 4.4 there
is an error. The fix is backwards compatible with earlier versions of
GCC.

Note, I'm using "GCC" to mean Gnu Compile Collection, as this problem is
specific to the C++ compiler, not the C compiler.