Add phone_type_choice grammar to Srec
Bug#232200
This grammar accepts the four phone types supported by the
VoiceDialer: "home", "mobile", "work", and "other", along
with "retry" and "exit".  It is used when the voicedialer
has determined which person to call, but is not yet sure
which phone to call them on.
Also get rid of some of the uneeded code from the boolean
grammar.  It was allowing for changes to be made at runtime,
but that is not needed.