Revert change Ibef3b834
Change Ibef3b834 increased the size of buffers
to address Bug: 2537307.  This was not an ideal
solution, because it consumes more memory and
is still vulnerable to failure if package names
or class names get larger.  Change I127c860b
fixes the problem in a different way, so we
can now shrink the buffer size back to the
original value.

Change-Id: I67db90017bb84b2c4a4ff962e5a524a9a32b67e8