Marking file-local inline functions as static
Compiling with clang results with undefined symbol errors at link time
for functions marked inline but not marked static or extern. A look at
the C spec indicates the compiler is free to assume the function exists
as a symbol elsewhere if static/extern are not used.  Since this
function really is meant to have static linkage, applied the static
keyword and the build is now fine.

Change-Id: I7c279c2592fe19ceda57fa3ecd565cc6f323b307
Author: Edwin Vane <edwin.vane@intel.com>
Reviewed-by: Kevin P Schoedel <kevin.p.schoedel@intel.com>