Update patch set 2
Patch Set 2:

Re: removal, I can chime in. As gkasten@ says, this is used *only* by the AOSP voice dialer (packages/apps/VoiceDialer)

This code is effectively dead. It's a snapshot of srec/ technology that's several years old that noone maintains (or understands even). It's effectively unchangeable because the tools to generate the voice input grammars etc. are not a part of this distribution.

enh@: Are you sure you saw libSR and not libRS ? I do see libSR_AudioIn on my N5 but that's only because it's listed as a required module in build/product/base.mk. I'll go delete that unnecessary rule now.

Patch-set: 2